Update Analyze.R
add plotit, calculate number of adults, adjust proportion of migrants so that it is correct.

consider if the initialized pop should be included as year 0

it is a nice day, somewhat chilly, and is the last day before Thanksgiving break. I am going to Michigan so that is exciting. Have a happy holiday and eat good food with people you love. this will be here when I get back. cheers!