ReproSucc troubleshooting (still)
ugh x2
getting frustrated that I havnt figured out how to even subset moms and dads yet but I know several ways NOT to!
also moms house is cold so there's that. no snow but *almost* blueish skies today with whispy clouds. Picked Christian's brain about going into academics. I'll cross that bridge when I get to it. fingers crossed for my future, and for the future of this ABM!