Add unique migrant SNPs
today I have added in the unique migrant and non-migrant SNPs. I have written it in such a way that I can put in a different frequency of alleles. for now, I am having fixed alleles in each of the source and focal pops.
today I am working from home for half the day because I am going dress shopping for Stephanie and Dane's wedding! It is a nice day, squirrels, birds, deer, and a cat have visited the house so far. Char is behind me in my office and Euchre is cuddled up at my feet in the blanket that is on my lap. Happy Thursday