Update RepSucc.R
finalize RepSucc.R since I forgot to finish the calculations LOL. proud of myself for this one really!

it is a nice sunny day. about to go to Con Gen class. then listen to seminar. was just on the phone with Mom. I miss her. summer is coming up quick though!