Update RunModel.R
added migSNPs to pop and source AND the proportion of migrant SNPs represented in the indv in the pop matrix.. edited break/next checks to add Analyze.R so that it doesn't keep skipping it. hopefully this is a solution and won't fuck anything up by putting a function in multiple times!

today is the last day of the Fall 2022 semester! I will be taking my end of semester photo on Monday after my end of semester meeting.

This semester was ROUGH, 100% due to my personal life, but unfortunately this ABM is moving slowly because of it. that is ok though. some balls will always be dropped. I'm still proud of you (aka me <3)