update to include migrants in repsuc
add migrants into reproductive success by changing the year born to the year they entered the population. there are still measures for just migrants, just natives, and all indv.

today is the Friday on the first day of the spring semester. it is rainy/overcast today. I am feeling ok. I am proud of all that I have accomplished on this, and I am excited to get it complete. The hardest part has been tying it all up at the end to make sure I have all the measures necessary. the next hard step will be cleaning it all up! but *I can do it*. Happy Friday the 13th!