Create FitBreed.R
this will sort generated babies in order of highest Ho so that high Ho babies are preferentially chosen. note that this can also be used to prefer indv with highest proportion of migrant alleles!!

**DONT FORGET that to use this function, in RunModel.R, change Breed to FitBreed

gotta say -- I am PROUD of myself. this took *relatively* no time at all to make and integrate into the script. I have come a LONGGG way both in knowing how to read and write code. I have learned so much and gotten SO MUCH BETTER at it!!