Add Ne_counts.csv
the table is created in three locations -- Breed.R, FitBreed.R, and RunModel.R. this is to track Ne and Nc to see if we can do something with these data. Note that if I want it to be more efficient, it is probs better to set up its own function and put the function in the multiple locations but that can be a problem for another day. think about the implications of putting a function within a function, as that would be necessary