adjust scripts after moving to easley
note the purge_mutants switch isn't working yet and breaks the simulation run. I believe the issue is in subsetting and then adding indv back into pop, most likely in FitnessDeath.R but might also be a problem in AgeDeath.R as well. Consider if this is even a switch that's necessary?